AITool.CSharp.Practice/Samples/4_AutoGen/python/requirements.txt
# AutoGen 相關套件
autogen-agentchat
autogen-ext[openai]

# 環境變數管理
python-dotenv